Kapua Freezes if Tag Deleted From Device fix

+++ console/module/tag/src/main/java/org/eclipse/kapua/app/console/module/tag/client/TagGrid.java
@@ -49,13 +49,15 @@ public class TagGrid extends EntityGrid<GwtTag> {
     @Override
     protected void selectionChangedEvent(GwtTag selectedItem) {
         super.selectionChangedEvent(selectedItem);
-        if (selectedItem != null) {
-            getToolbar().getEditEntityButton().setEnabled(currentSession.hasPermission(TagSessionPermission.write()));
-            getToolbar().getAddEntityButton().setEnabled(currentSession.hasPermission(TagSessionPermission.write()));
-            getToolbar().getDeleteEntityButton().setEnabled(currentSession.hasPermission(TagSessionPermission.delete()));
-        } else {
-            getToolbar().getEditEntityButton().setEnabled(false);
-            getToolbar().getDeleteEntityButton().setEnabled(false);
+        if (getToolbar().getEditEntityButton() != null && getToolbar().getAddEntityButton() != null && getToolbar().getDeleteEntityButton() != null) {
+            if (selectedItem != null) {
+                getToolbar().getEditEntityButton().setEnabled(currentSession.hasPermission(TagSessionPermission.write()));
+                getToolbar().getAddEntityButton().setEnabled(currentSession.hasPermission(TagSessionPermission.write()));
+                getToolbar().getDeleteEntityButton().setEnabled(currentSession.hasPermission(TagSessionPermission.delete()));
+            } else {
+                getToolbar().getEditEntityButton().setEnabled(false);
+                getToolbar().getDeleteEntityButton().setEnabled(false);
+            }
         }
     }
